Add eigensolvers & weighted products; upgrade Spectra to 1.0.1; bug fixes
Added
bd_eigs_sym() and bd_eigs() to compute top-k eigenvalues/eigenvectors (symmetric and general).
bd_wproduct() for in-memory weighted products (Xw, wX, Xᵀ w X, X w Xᵀ).
bdScalarwproduct() for in-memory scalar–matrix products.

Changed
Upgraded Spectra to 1.0.1 (from 0.9.x) and adapted code to the new API.

Fixed
Edge cases in apply helpers.
Cholesky decomposition failures on borderline (near-PSD) inputs.

Docs
Added roxygen/doxygen documentation for the new functions.

Notes
No user-facing breaking changes; developers need Spectra ≥ 1.0.1 to build.

+++ inst/include/Utilities/openme-utils.hpp
@@ -286,7 +286,7 @@ inline SEXP getDTthreads_R(SEXP verbose) {
         }
         
         if(threads.isNotNull()) {
-            if (Rcpp::as<int> (threads) <= ithreads){
+            if (Rcpp::as<int> (threads) <= (int)ithreads){
                 ithreads = Rcpp::as<int> (threads);
             }
         } else {
